Add some flags from regular cat; handle multiflags
added some of the flags on the man page of my ubuntu version of cat

in implementing all of these flags, it was easier to use std::getline
instead of getc and putc
this required some changes around file opening, and the handling of
stdin

also added the feature that single hyphen flags can be combined together
into a large group such as -ft, -Ebtfns, etc.
I thought about using getoptlong, but it seems to be more of a pain than
it's worth
one could use a vector to set up the array of struct options, but the
repetition needed to double check the args coming in is annoying.

also moved the usage print to a separate function because I wanted to
print the usage in a second spot when a bad hyphen flag is specified